double changed to decimal + generateTickets
tickets not correctly generated yet

+++ TicketShop/TicketShop/ISellable.cs
@@ -8,6 +8,6 @@ namespace TicketShop
 {
     interface ISellable
     {
-        double Price { get; }
+        decimal Price { get; }
     }
 }
